+++ packages/common/src/content/_internal/ContentSchema.ts
@@ -5,7 +5,8 @@ export type ContentEditorType = (typeof ContentEditorTypes)[number];
 export const ContentEditorTypes = [
   "hub:content:edit",
   "hub:content:settings",
-  "hub:content:discussions",
+  "hub:content:settings:discussions",
+  "hub:content:settings:discussions:compact",
 ] as const;
 
 /**