+++ packages/common/src/discussions/_internal/DiscussionSchema.ts
@@ -22,23 +22,5 @@ export const DiscussionSchema: IConfigurationSchema = {
       default: "",
       maxLength: 150,
     },
-    // TODO: externalize & spread onto HubItemEntitySchema rather than here
-    discussionSettings: {
-      type: "object",
-      properties: {
-        allowedChannelIds: {
-          type: "array",
-          items: {
-            type: "string",
-          },
-        },
-        allowedLocations: {
-          type: "array",
-          items: {
-            type: "object",
-          },
-        },
-      },
-    },
   },
 } as IConfigurationSchema;