+++ packages/common/src/items/slugs.ts
@@ -162,6 +162,7 @@ export function getUniqueSlug(
         : getUniqueSlug(slugInfo, requestOptions, step + 1)
     )
     .catch((e) => {
+      // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
       throw Error(`Error in getUniqueSlug ${e}`);
     });
 }