+++ packages/common/test/content/edit.test.ts
@@ -4,13 +4,15 @@ import * as adminModule from "@esri/arcgis-rest-feature-service";
 import { MOCK_AUTH, MOCK_HUB_REQOPTS, TOMORROW } from "../mocks/mock-auth";
 import * as modelUtils from "../../src/models";
 import { IModel } from "../../src/hub-types";
-import { IHubEditableContent } from "../../src/core/types";
+import { IGeometryInstance, IHubEditableContent } from "../../src/core/types";
 import {
   createContent,
   deleteContent,
   updateContent,
 } from "../../src/content/edit";
 import { cloneObject } from "../../src/util";
+import * as getDefaultEntitySettingsUtils from "../../src/discussions/api/settings/getDefaultEntitySettings";
+import * as createOrUpdateEntitySettingsUtils from "../../src/core/_internal/createOrUpdateEntitySettings";
 
 const GUID = "9b77674e43cf4bbd9ecad5189b3f1fdc";
 const myMockAuth = {
@@ -24,14 +26,18 @@ const myMockAuth = {
   username: "casey",
   password: "123456",
   portal: MOCK_HUB_REQOPTS.hubApiUrl,
+  getToken: () => Promise.resolve("fake-token"),
 } as any;
+const geometry: Partial<IGeometryInstance> = {
+  type: "polygon",
+  spatialReference: { wkid: 4326 },
+  toJSON: () => null,
+};
+export const DEFAULT_SETTINGS =
+  getDefaultEntitySettingsUtils.getDefaultEntitySettings("content");
 
 describe("content editing:", () => {
-  beforeAll(() => {
-    // suppress deprecation warnings
-    // tslint:disable-next-line: no-empty
-    spyOn(console, "warn").and.callFake(() => {});
-  });
+  let createOrUpdateEntitySettingsSpy: jasmine.Spy;
   describe("create content:", () => {
     it("converts to a model and creates the item", async () => {
       const createSpy = spyOn(modelUtils, "createModel").and.callFake(
@@ -41,9 +47,25 @@ describe("content editing:", () => {
           return Promise.resolve(newModel);
         }
       );
+      createOrUpdateEntitySettingsSpy = spyOn(
+        createOrUpdateEntitySettingsUtils,
+        "createOrUpdateEntitySettings"
+      ).and.returnValue(
+        Promise.resolve({
+          id: GUID,
+          ...DEFAULT_SETTINGS,
+        })
+      );
       const chk = await createContent(
-        { name: "Hello World", orgUrlKey: "dcdev" },
-        { authentication: MOCK_AUTH }
+        {
+          name: "Hello World",
+          orgUrlKey: "dcdev",
+          type: "Web Map",
+        },
+        {
+          ...MOCK_HUB_REQOPTS,
+          authentication: myMockAuth,
+        }
       );
 
       expect(chk.id).toBe(GUID);
@@ -54,6 +76,7 @@ describe("content editing:", () => {
       expect(modelToCreate.item.title).toBe("Hello World");
       // expect(modelToCreate.item.type).toBe("Hub Content");
       expect(modelToCreate.item.properties.orgUrlKey).toBe("dcdev");
+      expect(createOrUpdateEntitySettingsSpy.calls.count()).toBe(1);
     });
   });
   describe("update content:", () => {
@@ -61,6 +84,8 @@ describe("content editing:", () => {
     let getServiceSpy: jasmine.Spy;
     let updateModelSpy: jasmine.Spy;
     let updateServiceSpy: jasmine.Spy;
+    let createOrUpdateEntitySettingsSpy: jasmine.Spy;
+
     beforeEach(() => {
       getItemSpy = spyOn(portalModule, "getItem").and.returnValue(
         Promise.resolve({
@@ -81,12 +106,22 @@ describe("content editing:", () => {
       ).and.callFake((_url: string, opts: any) =>
         Promise.resolve(opts.updateDefinition)
       );
+      createOrUpdateEntitySettingsSpy = spyOn(
+        createOrUpdateEntitySettingsUtils,
+        "createOrUpdateEntitySettings"
+      ).and.returnValue(
+        Promise.resolve({
+          id: GUID,
+          ...DEFAULT_SETTINGS,
+        })
+      );
     });
     afterEach(() => {
       getItemSpy.calls.reset();
       getServiceSpy.calls.reset();
       updateModelSpy.calls.reset();
       updateServiceSpy.calls.reset();
+      createOrUpdateEntitySettingsSpy.calls.reset();
     });
     it("converts to a model and updates the item", async () => {
       const content: IHubEditableContent = {
@@ -126,6 +161,7 @@ describe("content editing:", () => {
       // No service is associated with Hub Initiatives
       expect(getServiceSpy).not.toHaveBeenCalled();
       expect(updateServiceSpy).not.toHaveBeenCalled();
+      expect(createOrUpdateEntitySettingsSpy.calls.count()).toBe(1);
     });
     it("handles when a location is explicitly set", async () => {
       const content: IHubEditableContent = {
@@ -319,6 +355,81 @@ describe("content editing:", () => {
       expect(chk.extendedProps?.downloads).toEqual(
         content.extendedProps?.downloads
       );
+    });
+    it("creates settings if none exists", async () => {
+      const content: IHubEditableContent = {
+        itemControl: "edit",
+        id: GUID,
+        name: "Hello World",
+        tags: ["Transportation"],
+        description: "Some longer description",
+        slug: "dcdev-wat-blarg",
+        orgUrlKey: "dcdev",
+        owner: "dcdev_dude",
+        type: "Hub Initiative",
+        createdDate: new Date(1595878748000),
+        createdDateSource: "item.created",
+        updatedDate: new Date(1595878750000),
+        updatedDateSource: "item.modified",
+        thumbnailUrl: "",
+        permissions: [],
+        schemaVersion: 1,
+        canEdit: false,
+        canDelete: false,
+        location: { type: "none" },
+        licenseInfo: "",
+        schedule: { mode: "automatic" },
+      };
+      const chk = await updateContent(content, {
+        ...MOCK_HUB_REQOPTS,
+        authentication: myMockAuth,
+      });
+      expect(chk.entitySettingsId).toBe(GUID);
+      expect(chk.discussionSettings).toEqual(
+        DEFAULT_SETTINGS.settings?.discussions
+      );
+    });
+    it("updates settings when settings id exists", async () => {
+      const content: IHubEditableContent = {
+        itemControl: "edit",
+        id: GUID,
+        name: "Hello World",
+        tags: ["Transportation"],
+        description: "Some longer description",
+        slug: "dcdev-wat-blarg",
+        orgUrlKey: "dcdev",
+        owner: "dcdev_dude",
+        type: "Hub Initiative",
+        createdDate: new Date(1595878748000),
+        createdDateSource: "item.created",
+        updatedDate: new Date(1595878750000),
+        updatedDateSource: "item.modified",
+        thumbnailUrl: "",
+        permissions: [],
+        schemaVersion: 1,
+        canEdit: false,
+        canDelete: false,
+        licenseInfo: "",
+        schedule: { mode: "automatic" },
+        entitySettingsId: "mock entity settings id",
+        location: {
+          type: "custom",
+          geometries: [geometry],
+        },
+        discussionSettings: {
+          allowedChannelIds: ["c1"],
+          allowedLocations: null,
+        },
+      };
+      const chk = await updateContent(content, {
+        ...MOCK_HUB_REQOPTS,
+        authentication: myMockAuth,
+      });
+      expect(chk.entitySettingsId).toBe(GUID);
+      expect(chk.discussionSettings).toEqual(
+        DEFAULT_SETTINGS.settings?.discussions
+      );
+      expect(createOrUpdateEntitySettingsSpy.calls.count()).toBe(1);
     });
   });
   describe("delete content", () => {