+++ packages/common/test/discussions/edit.test.ts
@@ -13,9 +13,9 @@ import { cloneObject } from "../../src/util";
 import * as settingUtils from "../../src/discussions/api/settings/settings";
 import * as getDefaultEntitySettingsUtils from "../../src/discussions/api/settings/getDefaultEntitySettings";
 import * as terraformer from "@terraformer/arcgis";
-import { EntitySettingType } from "../../src/discussions/api/types";
 import { Polygon } from "geojson";
 import { IGeometryInstance } from "../../src/core/types/IGeometryInstance";
+import * as createOrUpdateEntitySettingsUtils from "../../src/core/_internal/createOrUpdateEntitySettings";
 
 // TODO: update
 const GUID = "9b77674e43cf4bbd9ecad5189b3f1fdc";
@@ -63,6 +63,7 @@ const geometryTransformed: Polygon = {
 };
 
 describe("discussions edit:", () => {
+  let createOrUpdateEntitySettingsSpy: jasmine.Spy;
   describe("deleteDiscussion:", () => {
     it("deletes the item", async () => {
       const removeItemSpy = spyOn(portalModule, "removeItem").and.returnValue(
@@ -91,6 +92,20 @@ describe("discussions edit:", () => {
   });
 
   describe("createDiscussion:", () => {
+    beforeEach(() => {
+      createOrUpdateEntitySettingsSpy = spyOn(
+        createOrUpdateEntitySettingsUtils,
+        "createOrUpdateEntitySettings"
+      ).and.returnValue(
+        Promise.resolve({
+          id: GUID,
+          ...DEFAULT_SETTINGS,
+        })
+      );
+    });
+    afterEach(() => {
+      createOrUpdateEntitySettingsSpy.calls.reset();
+    });
     it("works with very limited structure", async () => {
       const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
         Promise.resolve("dcdev|hello-world")
@@ -101,15 +116,6 @@ describe("discussions edit:", () => {
           newModel.item.id = GUID;
           return Promise.resolve(newModel);
         }
-      );
-      const createSettingsSpy = spyOn(
-        settingUtils,
-        "createSettingV2"
-      ).and.returnValue(
-        Promise.resolve({
-          id: GUID,
-          ...DEFAULT_SETTINGS,
-        })
       );
       const chk = await createDiscussion(
         { name: "Hello World", orgUrlKey: "dcdev" },
@@ -128,27 +134,12 @@ describe("discussions edit:", () => {
         { slug: "dcdev|hello-world" },
         "should recieve slug"
       );
-      // should create the item
-      expect(createSpy.calls.count()).toBe(1);
       const modelToCreate = createSpy.calls.argsFor(0)[0];
-      expect(createSettingsSpy).toHaveBeenCalledTimes(1);
-      expect(createSettingsSpy).toHaveBeenCalledWith({
-        data: {
-          id: GUID,
-          type: EntitySettingType.CONTENT,
-          settings: {
-            discussions: {
-              allowedChannelIds: null,
-              allowedLocations: null,
-            },
-          },
-        },
-        ...requestOptions,
-      });
       expect(modelToCreate.item.title).toBe("Hello World");
       expect(modelToCreate.item.type).toBe("Discussion");
       expect(modelToCreate.item.properties.slug).toBe("dcdev|hello-world");
       expect(modelToCreate.item.properties.orgUrlKey).toBe("dcdev");
+      expect(createOrUpdateEntitySettingsSpy).toHaveBeenCalledTimes(1);
     });
     it("works with more complete object", async () => {
       // Note: this covers a branch when a slug is passed in
@@ -161,15 +152,6 @@ describe("discussions edit:", () => {
           newModel.item.id = GUID;
           return Promise.resolve(newModel);
         }
-      );
-      const createSettingsSpy = spyOn(
-        settingUtils,
-        "createSettingV2"
-      ).and.returnValue(
-        Promise.resolve({
-          id: GUID,
-          ...DEFAULT_SETTINGS,
-        })
       );
       const chk = await createDiscussion(
         {
@@ -197,25 +179,10 @@ describe("discussions edit:", () => {
         { slug: "dcdev|hello-world" },
         "should recieve slug"
       );
-      // should create the item
-      expect(createSpy.calls.count()).toBe(1);
       const modelToCreate = createSpy.calls.argsFor(0)[0];
-      expect(createSettingsSpy).toHaveBeenCalledTimes(1);
-      expect(createSettingsSpy).toHaveBeenCalledWith({
-        data: {
-          id: GUID,
-          type: EntitySettingType.CONTENT,
-          settings: {
-            discussions: {
-              allowedChannelIds: ["c1"],
-              allowedLocations: [geometryTransformed],
-            },
-          },
-        },
-        ...requestOptions,
-      });
       expect(modelToCreate.item.properties.slug).toBe("dcdev|hello-world");
       expect(modelToCreate.item.properties.orgUrlKey).toBe("dcdev");
+      expect(createOrUpdateEntitySettingsSpy).toHaveBeenCalledTimes(1);
     });
     it("coerces allowedChannelIds and allowedLocations to null when they are empty arrays", async () => {
       // Note: this covers a branch when a slug is passed in
@@ -228,15 +195,6 @@ describe("discussions edit:", () => {
           newModel.item.id = GUID;
           return Promise.resolve(newModel);
         }
-      );
-      const createSettingsSpy = spyOn(
-        settingUtils,
-        "createSettingV2"
-      ).and.returnValue(
-        Promise.resolve({
-          id: GUID,
-          ...DEFAULT_SETTINGS,
-        })
       );
       const chk = await createDiscussion(
         {
@@ -264,29 +222,28 @@ describe("discussions edit:", () => {
         { slug: "dcdev|hello-world" },
         "should recieve slug"
       );
-      // should create the item
-      expect(createSpy.calls.count()).toBe(1);
       const modelToCreate = createSpy.calls.argsFor(0)[0];
-      expect(createSettingsSpy).toHaveBeenCalledTimes(1);
-      expect(createSettingsSpy).toHaveBeenCalledWith({
-        data: {
-          id: GUID,
-          type: EntitySettingType.CONTENT,
-          settings: {
-            discussions: {
-              allowedChannelIds: null,
-              allowedLocations: null,
-            },
-          },
-        },
-        ...requestOptions,
-      });
       expect(modelToCreate.item.properties.slug).toBe("dcdev|hello-world");
       expect(modelToCreate.item.properties.orgUrlKey).toBe("dcdev");
+      expect(createOrUpdateEntitySettingsSpy).toHaveBeenCalledTimes(1);
     });
   });
 
   describe("updateDiscussion: ", () => {
+    beforeEach(() => {
+      createOrUpdateEntitySettingsSpy = spyOn(
+        createOrUpdateEntitySettingsUtils,
+        "createOrUpdateEntitySettings"
+      ).and.returnValue(
+        Promise.resolve({
+          id: GUID,
+          ...DEFAULT_SETTINGS,
+        })
+      );
+    });
+    afterEach(() => {
+      createOrUpdateEntitySettingsSpy.calls.reset();
+    });
     it("updates backing model and creates settings if none exist", async () => {
       const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
         Promise.resolve("dcdev|dcdev-wat-blarg-1")
@@ -298,15 +255,6 @@ describe("discussions edit:", () => {
         (m: IModel) => {
           return Promise.resolve(m);
         }
-      );
-      const createSettingsSpy = spyOn(
-        settingUtils,
-        "createSettingV2"
-      ).and.returnValue(
-        Promise.resolve({
-          id: GUID,
-          ...DEFAULT_SETTINGS,
-        })
       );
       const disc: IHubDiscussion = {
         itemControl: "edit",
@@ -349,24 +297,11 @@ describe("discussions edit:", () => {
       expect(getModelSpy.calls.count()).toBe(1);
       expect(updateModelSpy.calls.count()).toBe(1);
       const modelToUpdate = updateModelSpy.calls.argsFor(0)[0];
-      expect(createSettingsSpy).toHaveBeenCalledTimes(1);
-      expect(createSettingsSpy).toHaveBeenCalledWith({
-        data: {
-          id: GUID,
-          type: EntitySettingType.CONTENT,
-          settings: {
-            discussions: {
-              allowedChannelIds: null,
-              allowedLocations: null,
-            },
-          },
-        },
-        ...requestOptions,
-      });
       expect(modelToUpdate.item.description).toBe(disc.description);
       expect(modelToUpdate.item.properties.slug).toBe(
         "dcdev|dcdev-wat-blarg-1"
       );
+      expect(createOrUpdateEntitySettingsSpy).toHaveBeenCalledTimes(1);
     });
     it("updates backing model and updates settings when settings id exists", async () => {
       const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
@@ -379,15 +314,6 @@ describe("discussions edit:", () => {
         (m: IModel) => {
           return Promise.resolve(m);
         }
-      );
-      const updateSettingsSpy = spyOn(
-        settingUtils,
-        "updateSettingV2"
-      ).and.returnValue(
-        Promise.resolve({
-          id: GUID,
-          ...DEFAULT_SETTINGS,
-        })
       );
       spyOn(terraformer, "arcgisToGeoJSON").and.returnValue(
         geometryTransformed
@@ -441,24 +367,12 @@ describe("discussions edit:", () => {
       );
       expect(getModelSpy.calls.count()).toBe(1);
       expect(updateModelSpy.calls.count()).toBe(1);
-      expect(updateSettingsSpy).toHaveBeenCalledTimes(1);
-      expect(updateSettingsSpy).toHaveBeenCalledWith({
-        id: disc.entitySettingsId,
-        data: {
-          settings: {
-            discussions: {
-              allowedChannelIds: ["c1"],
-              allowedLocations: [geometryTransformed],
-            },
-          },
-        },
-        ...requestOptions,
-      });
       const modelToUpdate = updateModelSpy.calls.argsFor(0)[0];
       expect(modelToUpdate.item.description).toBe(disc.description);
       expect(modelToUpdate.item.properties.slug).toBe(
         "dcdev|dcdev-wat-blarg-1"
       );
+      expect(createOrUpdateEntitySettingsSpy).toHaveBeenCalledTimes(1);
     });
     describe("allowedLocations", () => {
       it("processes locations and persists them to settings", async () => {
@@ -477,15 +391,6 @@ describe("discussions edit:", () => {
           terraformer,
           "arcgisToGeoJSON"
         ).and.returnValue(geometryTransformed);
-        const updateSettingsSpy = spyOn(
-          settingUtils,
-          "updateSettingV2"
-        ).and.returnValue(
-          Promise.resolve({
-            id: GUID,
-            ...DEFAULT_SETTINGS,
-          })
-        );
         const disc: IHubDiscussion = {
           itemControl: "edit",
           id: GUID,
@@ -531,7 +436,7 @@ describe("discussions edit:", () => {
         );
         expect(getModelSpy.calls.count()).toBe(1);
         expect(updateModelSpy.calls.count()).toBe(1);
-        expect(updateSettingsSpy.calls.count()).toBe(1);
+        expect(createOrUpdateEntitySettingsSpy.calls.count()).toBe(1);
         const modelToUpdate = updateModelSpy.calls.argsFor(0)[0];
         expect(modelToUpdate.item.description).toBe(disc.description);
         expect(modelToUpdate.item.properties.slug).toBe(
@@ -539,10 +444,9 @@ describe("discussions edit:", () => {
         );
         // should transform location and persist to settings
         expect(arcgisToGeoJSONSpy.calls.count()).toBe(1);
-        const settingsToUpdate = updateSettingsSpy.calls.argsFor(0)[0];
-        expect(
-          settingsToUpdate.data.settings.discussions.allowedLocations
-        ).toEqual([geometryTransformed]);
+        const settingsToUpdate =
+          createOrUpdateEntitySettingsSpy.calls.argsFor(0)[2];
+        expect(settingsToUpdate).toEqual([geometryTransformed]);
       });
       it("processes locations of type = 'none' and persists them to settings", async () => {
         const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
@@ -560,15 +464,6 @@ describe("discussions edit:", () => {
           terraformer,
           "arcgisToGeoJSON"
         ).and.returnValue(geometryTransformed);
-        const updateSettingsSpy = spyOn(
-          settingUtils,
-          "updateSettingV2"
-        ).and.returnValue(
-          Promise.resolve({
-            id: GUID,
-            ...DEFAULT_SETTINGS,
-          })
-        );
         const disc: IHubDiscussion = {
           itemControl: "edit",
           id: GUID,
@@ -613,7 +508,7 @@ describe("discussions edit:", () => {
         );
         expect(getModelSpy.calls.count()).toBe(1);
         expect(updateModelSpy.calls.count()).toBe(1);
-        expect(updateSettingsSpy.calls.count()).toBe(1);
+        expect(createOrUpdateEntitySettingsSpy.calls.count()).toBe(1);
         const modelToUpdate = updateModelSpy.calls.argsFor(0)[0];
         expect(modelToUpdate.item.description).toBe(disc.description);
         expect(modelToUpdate.item.properties.slug).toBe(
@@ -621,10 +516,9 @@ describe("discussions edit:", () => {
         );
         // should transform location and persist to settings
         expect(arcgisToGeoJSONSpy.calls.count()).toBe(0);
-        const settingsToUpdate = updateSettingsSpy.calls.argsFor(0)[0];
-        expect(
-          settingsToUpdate.data.settings.discussions.allowedLocations
-        ).toBeNull();
+        const settingsToUpdate =
+          createOrUpdateEntitySettingsSpy.calls.argsFor(0)[0];
+        expect(settingsToUpdate.discussionSettings.allowedLocations).toBeNull();
       });
       it("processes locations and handles error if thrown", async () => {
         const slugSpy = spyOn(slugUtils, "getUniqueSlug").and.returnValue(
@@ -645,15 +539,6 @@ describe("discussions edit:", () => {
         const consoleWarnSpy = spyOn(console, "warn").and.callFake(() => {
           return;
         });
-        const updateSettingsSpy = spyOn(
-          settingUtils,
-          "updateSettingV2"
-        ).and.returnValue(
-          Promise.resolve({
-            id: GUID,
-            ...DEFAULT_SETTINGS,
-          })
-        );
         const disc: IHubDiscussion = {
           itemControl: "edit",
           id: GUID,
@@ -699,7 +584,7 @@ describe("discussions edit:", () => {
         );
         expect(getModelSpy.calls.count()).toBe(1);
         expect(updateModelSpy.calls.count()).toBe(1);
-        expect(updateSettingsSpy.calls.count()).toBe(1);
+        expect(createOrUpdateEntitySettingsSpy.calls.count()).toBe(1);
         const modelToUpdate = updateModelSpy.calls.argsFor(0)[0];
         expect(modelToUpdate.item.description).toBe(disc.description);
         expect(modelToUpdate.item.properties.slug).toBe(
@@ -707,10 +592,11 @@ describe("discussions edit:", () => {
         );
         // should transform location and persist to settings
         expect(arcgisToGeoJSONSpy.calls.count()).toBe(1);
-        const settingsToUpdate = updateSettingsSpy.calls.argsFor(0)[0];
-        expect(
-          settingsToUpdate.data.settings.discussions.allowedLocations
-        ).toEqual(null);
+        const settingsToUpdate =
+          createOrUpdateEntitySettingsSpy.calls.argsFor(0)[0];
+        expect(settingsToUpdate.discussionSettings.allowedLocations).toEqual(
+          null
+        );
         expect(consoleWarnSpy.calls.count()).toBe(1);
       });
     });